Use long for the page counter in CLI.pviewchain (CodeQL java/comparison-with-wider-type)
The "pviewchain" CLI command walks the right-sibling chain of B-tree
pages and caps the number of pages it prints with a counter guard:

    long count = 0;
    while (currentPage > 0 && count++ < maxcount) { ... }

maxcount is declared as long, but count was an int. The counter is the
only safeguard against an unbounded walk over a corrupt (cyclic)
right-sibling chain. If count ever wrapped past Integer.MAX_VALUE it
would go negative and, promoted to long, stay below maxcount forever,
defeating the cap. Widening count to long keeps it monotonically
increasing and matches the type of maxcount, so the int-vs-long
comparison is gone and the safety cap holds.

For normal inputs (maxcount is bounded 1..1_000_000 by the @Arg spec)
behaviour is unchanged.

Fixes code-scanning alert #54.

+++ persistit/core/src/main/java/com/persistit/CLI.java
@@ -1,5 +1,6 @@
 /**
  * Copyright 2011-2012 Akiban Technologies, Inc.
+ * Portions Copyrighted 2026 3A Systems, LLC
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -1119,7 +1120,7 @@ public class CLI {
             @Override
             protected void runTask() throws Exception {
                 long currentPage = pageAddress;
-                int count = 0;
+                long count = 0;
                 while (currentPage > 0 && count++ < maxcount) {
                     if (_currentVolume == null) {
                         postMessage("Select a volume", LOG_NORMAL);
